Factor out libraries section
